+++ src/__tests__/example.test.ts
@@ -1,7 +1,3 @@
-/* eslint-disable no-console */
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-/* eslint-disable @typescript-eslint/no-unsafe-return */
 import { parseExpressApp } from '../index';
 import express, { RequestHandler, Request, Response, NextFunction } from 'express';